backend/requirements/dev.txt
# Include base requirements
-r base.txt

# Development tools
django-debug-toolbar>=4.2
ipython>=8.12
ipdb>=0.13

# Code quality
black>=23.0
flake8>=6.0
isort>=5.12

# Testing (for local dev)
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
factory-boy>=3.3
faker>=19.0
